Refactor block interactable logic to include a serializer
Multiple changes are made to the block interactable logic in AquaticSeriesLib. The main changes include the addition of a serializer and refactoring how blocks are tracked and spawned. The AbstractInteractable class now has a serializer which handles the serialization and deserialization of block states. Also, instead of tracking only the locations of blocks, we now additionally store an ID for each block. This shift from purely spacial tracking to a serialized model allows for more robust and flexible handling of block states.